Adding method to reload all item views + data.
Currently reloadData only reloads the grid based on the stored data, but recycled item views are not cleared. This is not useful if you need to switch the UIView that needs to be rendered in the grid. Because dequeueReusableItem does not provide an interface to dequeue items with a specific reuse identifier, this is the next cleanest solution.