Fixing layout re-rendering
When reloading views, all older views need to be removed by iterating through the _visibleItems array and calling the proper function. After that, we need to reload the data with YES as the parameter so the delegate methods get called for each cell. In addition, we force the view to relayout with layoutSubviews.