Improved API name for preloading
Previous api name was preloadBuffer, which wasn't intuitive. Renamed the property to preloadRowSpan to indicate to developers that the number assigned
to this will have x number of rows preloaded before the first visible row and x number of rows preloaded after the last visible row. preloadBuffer is still available, but has been marked as deprecated.