Support for setting a maximum row gap.
Normally the row gap is calculated to be the same as the column gap. But
for some grid views that need to be synchronized to a background, this
creates issues. Now the row gap is calculated by the lowest value
between maximumRowGap and column gap. By default, as long as the column
gap is less than the maximum row gap (default=INFINITY), then everything
works as normal. Basically, now you can set maximum row gap to zero to
support synchronized background tiling with a grid.